REFACTOR: Extract model/diagram manipulators into `document-types-ts`.
