+++ packages/document-types-ts/src/deepcopy.ts
@@ -1,0 +1,4 @@
+// FIXME: This shouldn't be here.
+export function deepCopyJSON(value: unknown) {
+    return JSON.parse(JSON.stringify(value));
+}
