+++ packages/document-types-ts/src/diagram.ts
@@ -1,7 +1,32 @@
 import { v7 } from "uuid";
 
-import type { DiagramJudgment, Mor, MorType, Ob, ObType } from "catlog-wasm";
-import { deepCopyJSON } from "../util/deepcopy";
+import { deepCopyJSON } from "./deepcopy";
+import { newNotebook } from "./notebook";
+import type {
+    Document,
+    DiagramJudgment,
+    Mor,
+    MorType,
+    Ob,
+    ObType,
+    StableRef,
+} from "./rs_document_types";
+import { currentVersion } from "./rs_document_types";
+
+/** A document defining a diagram in a model. */
+export type DiagramDocument = Document & { type: "diagram" };
+
+/** Create an empty diagram of a model. */
+export const newDiagramDocument = (modelRef: StableRef): DiagramDocument => ({
+    name: "",
+    type: "diagram",
+    diagramIn: {
+        ...modelRef,
+        type: "diagram-in",
+    },
+    notebook: newNotebook<DiagramJudgment>(),
+    version: currentVersion(),
+});
 
 /** Create a new diagram object declaration with the given object type. */
 export const newDiagramObjectDecl = (
