+++ packages/document-types-ts/src/index.ts
@@ -1,3 +1,9 @@
-export * from "./types";
+export * from "./rs_document_types";
 
+export type { DiagramDocument } from "./diagram";
+export type { ModelDocument } from "./model";
+export type { FormalCell, RichTextCell, StemCell } from "./notebook";
+
+export * as Diagram from "./diagram";
+export * as Model from "./model";
 export * as Nb from "./notebook";
