+++ packages/document-types-ts/src/rs_document_types.ts
@@ -1,0 +1,1 @@
+export * from "../dist/pkg-node/document_types";
