+++ packages/frontend/src/api/document_editing.test.ts
@@ -8,8 +8,7 @@ import invariant from "tiny-invariant";
 import { v4 } from "uuid";
 import { afterAll, assert, describe, test } from "vitest";
 
-import type { Document } from "catlog-wasm";
-import type { ModelDocument } from "../model/document";
+import type { Document, ModelDocument } from "document-types";
 import { normalizeImmutableStrings } from "../util/immutable_string";
 import { createTestDocument, initTestUserAuth } from "../util/test_util.ts";
 import { createFetchWithAuth, createRpcClient, unwrap } from "./rpc.ts";
