+++ packages/frontend/src/diagram/diagram_editor.tsx
@@ -2,8 +2,8 @@ import { MultiProvider } from "@solid-primitives/context";
 import { Match, Switch, useContext } from "solid-js";
 import invariant from "tiny-invariant";
 
-import type { DiagramJudgment, DiagramMorDecl, DiagramObDecl } from "catlog-wasm";
-import { Nb } from "document-types";
+import type { DiagramJudgment, DiagramMorDecl, DiagramObDecl } from "document-types";
+import { Diagram, Nb } from "document-types";
 import { LiveModelContext } from "../model";
 import { type CellConstructor, type FormalCellEditorProps, NotebookEditor } from "../notebook";
 import type { InstanceTypeMeta } from "../theory";
@@ -11,7 +11,6 @@ import { LiveDiagramContext } from "./context";
 import type { LiveDiagramDoc } from "./document";
 import { DiagramMorphismCellEditor } from "./morphism_cell_editor";
 import { DiagramObjectCellEditor } from "./object_cell_editor";
-import { duplicateDiagramJudgment, newDiagramMorphismDecl, newDiagramObjectDecl } from "./types";
 
 /** Notebook editor for a diagram in a model.
  */
@@ -39,7 +38,7 @@ export function DiagramNotebookEditor(props: { liveDiagram: LiveDiagramDoc }) {
                 formalCellEditor={DiagramCellEditor}
                 cellConstructors={cellConstructors()}
                 cellLabel={judgmentLabel}
-                duplicateCell={duplicateDiagramJudgment}
+                duplicateCell={Diagram.duplicateDiagramJudgment}
             />
         </MultiProvider>
     );
@@ -92,9 +91,9 @@ function diagramCellConstructor(meta: InstanceTypeMeta): CellConstructor<Diagram
         construct() {
             switch (tag) {
                 case "ObType":
-                    return Nb.newFormalCell(newDiagramObjectDecl(meta.obType));
+                    return Nb.newFormalCell(Diagram.newDiagramObjectDecl(meta.obType));
                 case "MorType":
-                    return Nb.newFormalCell(newDiagramMorphismDecl(meta.morType));
+                    return Nb.newFormalCell(Diagram.newDiagramMorphismDecl(meta.morType));
                 default:
                     throw tag satisfies never;
             }
