+++ packages/frontend/src/diagram/document.ts
@@ -1,33 +1,15 @@
 import type { AnyDocumentId, Repo } from "@automerge/automerge-repo";
 import { type Accessor, createMemo } from "solid-js";
 
-import type {
-    DblModelDiagram,
-    DiagramJudgment,
-    Document,
-    ModelDiagramValidationResult,
-    StableRef,
-    Uuid,
+import {
+    type DblModelDiagram,
+    elaborateDiagram,
+    type ModelDiagramValidationResult,
 } from "catlog-wasm";
-import { currentVersion, elaborateDiagram } from "catlog-wasm";
-import { Nb } from "document-types";
+import type { DiagramDocument, DiagramJudgment, StableRef, Uuid } from "document-types";
+import { Diagram, Nb } from "document-types";
 import { type Api, type DocRef, findAndMigrate, type LiveDoc, makeLiveDoc } from "../api";
 import type { LiveModelDoc, ModelLibrary } from "../model";
-
-/** A document defining a diagram in a model. */
-export type DiagramDocument = Document & { type: "diagram" };
-
-/** Create an empty diagram of a model. */
-export const newDiagramDocument = (modelRef: StableRef): DiagramDocument => ({
-    name: "",
-    type: "diagram",
-    diagramIn: {
-        ...modelRef,
-        type: "diagram-in",
-    },
-    notebook: Nb.newNotebook<DiagramJudgment>(),
-    version: currentVersion(),
-});
 
 /** A diagram document "live" for editing. */
 export type LiveDiagramDoc = {
@@ -125,7 +107,7 @@ export function enlivenDiagramDocument(
 
 /** Create a new, empty diagram in the backend. */
 export function createDiagram(api: Api, inModel: StableRef): Promise<string> {
-    const init = newDiagramDocument(inModel);
+    const init = Diagram.newDiagramDocument(inModel);
     return api.createDoc(init);
 }
