+++ packages/frontend/src/diagram/index.ts
@@ -1,3 +1,2 @@
 export * from "./context";
 export * from "./document";
-export * from "./types";
