+++ packages/frontend/src/model/index.ts
@@ -1,4 +1,4 @@
 export * from "./context";
 export * from "./document";
 export * from "./model_library";
-export * from "./types";
+export * from "./label";
